% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignMitogenomes.R
\name{alignMitogenomes}
\alias{alignMitogenomes}
\title{alignMitogenomes}
\usage{
alignMitogenomes(
  alignment.folder = NULL,
  genbank.file = NULL,
  draft.contigs = "draftContigs",
  output.dir = "Genomes",
  dataset.name = "untrimmed",
  overwrite = FALSE
)
}
\arguments{
\item{alignment.folder}{path to a folder of sequence alignments in phylip format.}

\item{genbank.file}{available input alignment formats: fasta or phylip}

\item{draft.contigs}{contigs are added into existing alignment if algorithm is "add"}

\item{output.dir}{available output formats: phylip}

\item{dataset.name}{remove samples too divergent from consensus, values 0-1 for proportion similar sites}

\item{overwrite}{TRUE to supress mafft screen output}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for batch trimming a folder of alignments, with the various trimming functions available to select from
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
