% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMitogenomes.R
\name{buildMitogenomes}
\alias{buildMitogenomes}
\title{buildMitogenomes}
\usage{
buildMitogenomes(
  annotation.dir = "Annotations",
  alignment.folder = "Alignments/untrimmed-alignments",
  genome.alignment = "Genomes/alignments/untrimmed_mitogenome_alignment.phy",
  genome.dir = "Genomes",
  output.dir = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{annotation.dir}{path to a folder of sequence alignments in phylip format.}

\item{alignment.folder}{path to a folder of sequence alignments in phylip format.}

\item{genome.alignment}{available input alignment formats: fasta or phylip}

\item{genome.dir}{contigs are added into existing alignment if algorithm is "add"}

\item{output.dir}{available output formats: phylip}

\item{overwrite}{TRUE to supress mafft screen output}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for batch trimming a folder of alignments, with the various trimming functions available to select from
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
