% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findORF.R
\name{findORF}
\alias{findORF}
\title{findORF}
\usage{
findORF(input.seq = NULL, codons = F, min.size = 80)
}
\arguments{
\item{input.seq}{path to a folder of sequence alignments in phylip format.}

\item{codons}{algorithm to use: "add" add sequences with "add.contigs"; "localpair" for local pair align. All others available}

\item{min.size}{TRUE to supress screen output}

\item{genbank.file}{contigs are added into existing alignment if algorithm is "add"}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for running the program spades to assemble short read sequencing data
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
