% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCircularGenome.R
\name{isCircularGenome}
\alias{isCircularGenome}
\title{isCircularGenome}
\usage{
isCircularGenome(contig = NULL, blast.path = "blastn", cap3.path = "cap3")
}
\arguments{
\item{contig}{path to a folder of raw reads in fastq format.}

\item{blast.path}{the new directory to save the adaptor trimmed sequences}

\item{genbank.reference}{a csv file with a "File" and "Sample" columns, where "File" is the file name and "Sample" is the desired renamed file}
}
\value{
a new directory of adaptor trimmed reads and a summary of the trimming in logs/
}
\description{
Function for removing adaptor sequences from raw Illumina sequence data using the program fastp
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
