% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReference.R
\name{makeReference}
\alias{makeReference}
\title{makeReference}
\usage{
makeReference(genbank.file = NULL, overwrite = TRUE, rep.origin = FALSE)
}
\arguments{
\item{genbank.file}{path to a folder of sequence alignments in phylip format.}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for running the program spades to assemble short read sequencing data
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
