% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOffTarget.R
\name{removeOffTarget}
\alias{removeOffTarget}
\title{removeOffTarget}
\usage{
removeOffTarget(
  target = NULL,
  contigs = NULL,
  blast.path = "blast",
  threads = 1,
  remove.bad = F,
  quiet = TRUE
)
}
\arguments{
\item{target}{path to a folder of sequence alignments in phylip format.}

\item{contigs}{contigs are added into existing alignment if algorithm is "add"}

\item{blast.path}{contigs are added into existing alignment if algorithm is "add"}

\item{threads}{algorithm to use: "add" add sequences with "add.contigs"; "localpair" for local pair align. All others available}

\item{quiet}{TRUE to supress screen output}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for running the program spades to assemble short read sequencing data
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
