% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimMitoAlignments.R
\name{trimMitoAlignments}
\alias{trimMitoAlignments}
\title{trimMitoAlignments}
\usage{
trimMitoAlignments(
  alignment.dir = "Alignments/untrimmed-alignments",
  alignment.format = "phylip",
  output.dir = "Alignments/trimmed-alignments",
  output.format = "phylip",
  sample.similiarity = TRUE,
  TrimAl = TRUE,
  TrimAl.path = "trimal",
  trim.external = TRUE,
  min.external.percent = 50,
  trim.coverage = TRUE,
  min.coverage.percent = 50,
  trim.column = TRUE,
  min.column.gap.percent = 100,
  alignment.assess = TRUE,
  min.sample.bp = 0,
  min.align.length = 0,
  min.taxa.count = 0,
  min.gap.percent = 0,
  overwrite = FALSE
)
}
\arguments{
\item{alignment.dir}{path to a folder of sequence alignments in phylip format.}

\item{alignment.format}{available input alignment formats: fasta or phylip}

\item{output.dir}{contigs are added into existing alignment if algorithm is "add"}

\item{output.format}{available output formats: phylip}

\item{TrimAl}{if a file name is provided, save.name will be used to save aligment to file as a fasta}

\item{TrimAl.path}{path to a folder of sequence alignments in phylip format.}

\item{trim.external}{give a save name if you wnat to save the summary to file.}

\item{min.external.percent}{TRUE to supress mafft screen output}

\item{trim.coverage}{path to a folder of sequence alignments in phylip format.}

\item{min.coverage.percent}{contigs are added into existing alignment if algorithm is "add"}

\item{trim.column}{algorithm to use: "add" add sequences with "add.contigs"; "localpair" for local pair align. All others available}

\item{min.column.gap.percent}{TRUE applies the adjust sequence direction function of MAFFT}

\item{alignment.assess}{if a file name is provided, save.name will be used to save aligment to file as a fasta}

\item{min.sample.bp}{path to a folder of sequence alignments in phylip format.}

\item{min.align.length}{give a save name if you wnat to save the summary to file.}

\item{min.taxa.count}{TRUE to supress mafft screen output}

\item{min.gap.percent}{if a file name is provided, save.name will be used to save aligment to file as a fasta}

\item{overwrite}{TRUE to supress mafft screen output}

\item{sample.similarity}{remove samples too divergent from consensus, values 0-1 for proportion similar sites}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for batch trimming a folder of alignments, with the various trimming functions available to select from
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
