% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNAscanBatch.R
\name{tRNAscanBatch}
\alias{tRNAscanBatch}
\title{tRNAscanBatch}
\usage{
tRNAscanBatch(
  genome.dir = NULL,
  genbank.file = NULL,
  out.dir = "tRNAscan",
  tRNAscan.path = "tRNAscan-SE",
  organism.type = c("mammal", "vertebrate", "eukaryotic"),
  overwrite = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{genome.dir}{path to a folder of sequence alignments in phylip format.}

\item{genbank.file}{contigs are added into existing alignment if algorithm is "add"}

\item{out.dir}{contigs are added into existing alignment if algorithm is "add"}

\item{tRNAscan.path}{algorithm to use: "add" add sequences with "add.contigs"; "localpair" for local pair align. All others available}

\item{organism.type}{algorithm to use: "add" add sequences with "add.contigs"; "localpair" for local pair align. All others available}

\item{overwrite}{TRUE to supress screen output}

\item{quiet}{TRUE to supress screen output}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for running the program spades to assemble short read sequencing data
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
